fastapi/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openai==2.6.0
pydantic==2.12.3
pydantic-settings==2.12.0
pydantic_core==2.41.4
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.21
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
sentry-sdk==2.48.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.49.3
tqdm==4.67.1
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
